Canonicalization of FloatArithmetic nodes

+++ dataset/GitHub_Java/oracle.graal/graal/com.oracle.max.graal.compiler/src/com/oracle/max/graal/compiler/ir/FloatAdd.java
@@ -83,7 +83,7 @@ public final class FloatAdd extends FloatArithmetic {
                         return x;
                     }
                 } else {
-                    assert kind == CiKind.Double;
+                    assert kind == CiKind.Long;
                     double c = y.asConstant().asDouble();
                     if (c == 0.0) {
                         return x;
